+++ js-api/ui.ts
@@ -32,7 +32,7 @@ import $ from 'cash-dom';
 import {__obs} from './src/events';
 import {HtmlUtils, _isDartium, _options} from './src/utils';
 import * as rxjs from 'rxjs';
-import { CanvasRenderer, GridCellRenderer, SemanticValue } from './src/grid';
+import {CanvasRenderer, GridCellRenderer, SemanticValue, Size} from './src/grid';
 import {Entity, FileInfo, Property, User} from './src/entities';
 import { Column, DataFrame } from './src/dataframe';
 import dayjs from "dayjs";
@@ -709,11 +709,6 @@ export namespace input {
     [key: string]: any;
   }
 
-  interface Size {
-    height: number;
-    width: number;
-  }
-
   const optionsMap: {[key: string]: (input: InputBase, option: any) => void} = {
     value: (input, x) => input.value = input.inputType === d4.InputType.File ? toDart(x) : x,
     nullable: (input, x) => input.nullable = x, // finish it?
@@ -739,6 +734,8 @@ export namespace input {
     const specificOptions = (({value, property, elementOptions, onCreated, onValueChanged, ...opt}) => opt)(options);
     if (!options.property)
       options.property = Property.fromOptions({name: input.caption, inputType: inputType as string});
+    if (specificOptions.nullable !== undefined)
+      optionsMap['nullable'](input, specificOptions.nullable);
     for (let key of Object.keys(specificOptions)) {
       if (['min', 'max', 'step', 'format', 'showSlider', 'showPlusMinus'].includes(key))
         (options.property as IIndexable)[key] = (specificOptions as IIndexable)[key];
@@ -747,7 +744,7 @@ export namespace input {
         inputType === d4.InputType.Column ? setColumnInputTable(input, (specificOptions as IIndexable)[key], filter) :
           setColumnsInputTable(input, (specificOptions as IIndexable)[key], filter);
       }
-      if (optionsMap[key] !== undefined)
+      if (key !== 'nullable' && optionsMap[key] !== undefined)
         optionsMap[key](input, (specificOptions as IIndexable)[key]);
     }
     const baseOptions = (({value, nullable, property, onCreated, onValueChanged}) => ({value, nullable, property, onCreated, onValueChanged}))(options);